+++ json-path/src/main/java/com/jayway/jsonpath/internal/spi/mapper/JacksonMappingProvider.java
@@ -14,17 +14,16 @@
  */
 package com.jayway.jsonpath.internal.spi.mapper;
 
+import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.jayway.jsonpath.Configuration;
+import com.jayway.jsonpath.TypeRef;
+import com.jayway.jsonpath.spi.mapper.MappingException;
 import com.jayway.jsonpath.spi.mapper.MappingProvider;
-
-import java.util.HashMap;
 
 public class JacksonMappingProvider implements MappingProvider {
 
     private final ObjectMapper objectMapper;
-
-    private HashMap<Class<?>, HashMap<Class<?>, Mapper>> converters = new HashMap<Class<?>, HashMap<Class<?>, Mapper>>();
 
     public JacksonMappingProvider() {
         this(new ObjectMapper());
@@ -33,14 +32,6 @@ public class JacksonMappingProvider implements MappingProvider {
     public JacksonMappingProvider(ObjectMapper objectMapper) {
         this.objectMapper = objectMapper;
     }
-    public void addMapper(Mapper converter) {
-        for (Mapper.ConvertiblePair convertible : converter.getConvertibleTypes()) {
-            if(!converters.containsKey(convertible.getTargetType())){
-                converters.put(convertible.getTargetType(), new HashMap<Class<?>, Mapper>());
-            }
-            converters.get(convertible.getTargetType()).put(convertible.getSourceType(), converter);
-        }
-    }
 
 
     @Override
@@ -48,13 +39,26 @@ public class JacksonMappingProvider implements MappingProvider {
         if(source == null){
             return null;
         }
-        HashMap<Class<?>, Mapper> targetConverters = converters.get(targetType);
-        if(targetConverters != null){
-            Mapper mapper = targetConverters.get(source.getClass());
-            if(mapper != null){
-                return (T) mapper.convert(source, source.getClass(), targetType, configuration);
-            }
+        try {
+            return objectMapper.convertValue(source, targetType);
+        } catch (Exception e) {
+            throw new MappingException(e);
         }
-        return objectMapper.convertValue(source, targetType);
+
+    }
+
+    @Override
+    public <T> T map(Object source, final TypeRef<T> targetType, Configuration configuration) {
+        if(source == null){
+            return null;
+        }
+        JavaType type = objectMapper.getTypeFactory().constructType(targetType.getType());
+
+        try {
+            return (T)objectMapper.convertValue(source, type);
+        } catch (Exception e) {
+            throw new MappingException(e);
+        }
+
     }
 }