+++ json-path/src/test/java/com/jayway/jsonpath/RemoveTest.java
@@ -3,6 +3,7 @@ package com.jayway.jsonpath;
 import static com.jayway.jsonpath.JsonPath.using;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -12,14 +13,16 @@ public class RemoveTest extends BaseTest {
 
     @Test
     public void fromJsonString() {
-    	String result = JsonPath.remove("{\"val\":1,\"val2\":2}", "$.val");
-    	assertThat(result).isEqualTo("{\"val2\":2}");
+    	Map<String, Object> result = JsonPath.delete("{\"val\":1,\"val2\":2}", "$.val");
+    	assertThat(result).isEqualTo(Collections.singletonMap("val2", 2));
     }
     
     @SuppressWarnings("unchecked")
 	@Test
     public void fromMap() {
-        Map<String, Object> model = new HashMap<String, Object>(){{
+        Map<String, Object> model = new HashMap<String, Object>(){
+			private static final long serialVersionUID = 1L;
+		{
             put("a", "a-val");
             put("b", "b-val");
             put("c", "c-val");
@@ -27,25 +30,9 @@ public class RemoveTest extends BaseTest {
 
         Configuration conf = Configuration.defaultConfiguration();
 
-        assertThat(using(conf).parse(model).remove("$.b", Map.class))
+        assertThat((Map<String, Object>)using(conf).parse(model).remove("$.b"))
                 .containsEntry("a", "a-val")
                 .containsEntry("c", "c-val").doesNotContainKey("b");
-    }
-    
-    @Test
-    public void fromMapToString() {
-        Map<String, Object> model = new HashMap<String, Object>(){{
-            put("a", "a-val");
-            put("b", "b-val");
-            put("c", "c-val");
-        }};
-
-        Configuration conf = Configuration.defaultConfiguration();
-
-        assertThat(using(conf).parse(model).remove("$.b", String.class))
-		        .contains("\"a\":\"a-val\"")
-		        .contains("\"c\":\"c-val\"")
-		        .doesNotContain("\"b\":\"b-val\"");
     }
     
     @SuppressWarnings("unchecked")
@@ -53,7 +40,7 @@ public class RemoveTest extends BaseTest {
     public void fromStringToMap() {
         Configuration conf = Configuration.defaultConfiguration();
 
-        assertThat(using(conf).parse("{\"val\":1,\"val2\":2}").remove("$.val2", Map.class))
+        assertThat((Map<String, Object>)using(conf).parse("{\"val\":1,\"val2\":2}").remove("$.val2"))
 		        .containsEntry("val", 1).doesNotContainKey("val2");
     }
 }